misc(native-pos): Extract AbstractNativeProcess from NativeExecutionProcess (#27696)
Summary:

## Description

Extracts the common subprocess lifecycle code from `NativeExecutionProcess` into a new `AbstractNativeProcess` base class. The original `NativeExecutionProcess` becomes a thin subclass that supplies execution-specific configuration. Behavior is unchanged.

## Motivation and Context

The presto-on-spark driver needs to launch a second native subprocess (a metadata-only sidecar — see follow-up diffs in this stack) that shares ~95% of its lifecycle with the existing per-task native worker process: spawn the binary, write per-process etc/ files, install a shutdown hook, monitor for unexpected exit. Without this refactor each new native subprocess type would have to re-implement the lifecycle boilerplate.

## Impact

No user-facing impact. `NativeExecutionProcess` is internal to presto-on-spark and its public behavior (constructor signature, start/stop semantics, configuration files written) is preserved. Subclasses can override hooks like `populateConfigurationFiles` and `resolveProcessWorkingPath`.

## Test Plan

Existing `NativeExecutionProcess` unit tests continue to pass. The new abstract class is exercised by tests in subsequent diffs in this stack that introduce a second concrete subclass.

## Contributor checklist

- [x] My PR adheres to the code style of this project.
- [x] My code builds clean without any errors or warnings.
- [x] I am willing to help maintain this change if there are any issues in the future.

## Release Notes

\`\`\`
== NO RELEASE NOTE ==
\`\`\`

Differential Revision: D103025709
